% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drawMapWithSites}
\alias{drawMapWithSites}
\title{Draw (France or...) map with all sites of colors 'cols'}
\usage{
drawMapWithSites(M, cols = rep(1, nrow(M)))
}
\arguments{
\item{M}{Coordinates matrix (in columns)}

\item{cols}{Vector of colors for each row of M [default: all black]}
}
\description{
Draw (France or...) map with all sites of colors 'cols'
}

