% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drawNeighborhoodGraph}
\alias{drawNeighborhoodGraph}
\title{Draw neighborhoods graph on top of a country map (or any other map)}
\usage{
drawNeighborhoodGraph(M, NI)
}
\arguments{
\item{M}{Coordinates matrix (in columns)}

\item{NI}{Neighborhoods of M rows (list of integer vectors)}
}
\description{
Draw neighborhoods graph on top of a country map (or any other map)
}

