% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCurves}
\alias{plotCurves}
\title{Plot a matrix of curves (in rows)}
\usage{
plotCurves(M, cols = rep(1, nrow(M)))
}
\arguments{
\item{M}{Coordinates matrix (in columns)}

\item{cols}{Vector of colors for each row of M [default: all black]}
}
\description{
Plot a matrix of curves (in rows)
}

